+++ tests/settings.test.ts
@@ -1,6 +1,6 @@
 import "ts-jest";
 import { IHighlightStyles, Logger, TLogLevelColor } from "../src";
-import { TUtilsInspectColors } from "../src/interfaces";
+import { doesLogContain } from "./helper";
 
 const stdOut = [];
 const stdErr = [];
@@ -162,6 +162,455 @@ describe("Logger: settings", () => {
     );
   });
 
+  /* date time */
+
+  /* dateTimePattern */
+  test("init logger: dateTimePattern: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    logger.info("test 123");
+
+    const month = new Date().getMonth() + 1;
+    const monthStr = month < 10 ? "0" + month : month.toString();
+    expect(
+      doesLogContain(
+        stdArray,
+        new Date().getFullYear().toString() + "-" + monthStr
+      )
+    ).toBeTruthy();
+  });
+
+  test("init logger: dateTimePattern: day.month.year", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      dateTimePattern: "day.month.year",
+    });
+    logger.info("test 123");
+
+    const month = new Date().getMonth() + 1;
+    const monthStr = month < 10 ? "0" + month : month.toString();
+    expect(
+      doesLogContain(
+        stdArray,
+        monthStr + "." + new Date().getFullYear().toString()
+      )
+    ).toBeTruthy();
+  });
+
+  /* dateTimeTimezone */
+  test("init logger: dateTimeTimezone: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    logger.info("test 123");
+
+    const [
+      ,
+      ,
+      ,
+      ,
+      ,
+      ,
+      { value: hour },
+      ,
+      { value: minute },
+    ] = new Intl.DateTimeFormat("en", {
+      weekday: undefined,
+      year: "numeric",
+      month: "2-digit",
+      day: "2-digit",
+      hour12: false,
+      hour: "2-digit",
+      minute: "2-digit",
+      second: "2-digit",
+      timeZone: "utc",
+    }).formatToParts(new Date());
+    expect(doesLogContain(stdArray, `${hour}:${minute}`)).toBeTruthy();
+  });
+
+  test("init logger: dateTimeTimezone: Europe/Berlin", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const timezone: string = "Europe/Berlin";
+    const logger: Logger = new Logger({
+      stdOut: std,
+      dateTimeTimezone: timezone,
+    });
+    logger.info("test 123");
+
+    const [
+      ,
+      ,
+      ,
+      ,
+      ,
+      ,
+      { value: hour },
+      ,
+      { value: minute },
+    ] = new Intl.DateTimeFormat("en", {
+      weekday: undefined,
+      year: "numeric",
+      month: "2-digit",
+      day: "2-digit",
+      hour12: false,
+      hour: "2-digit",
+      minute: "2-digit",
+      second: "2-digit",
+      timeZone: timezone,
+    }).formatToParts(new Date());
+    expect(doesLogContain(stdArray, `${hour}:${minute}`)).toBeTruthy();
+  });
+
+  /* printLogMessageInNewLine */
+  test("init logger: printLogMessageInNewLine: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    logger.info("test 123");
+    expect(stdArray[3]).toBe("  \t");
+  });
+
+  test("init logger: printLogMessageInNewLine: true", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      printLogMessageInNewLine: true,
+    });
+    logger.info("test 123");
+    expect(stdArray[3]).toBe("\n");
+  });
+
+  test("init logger: printLogMessageInNewLine: false", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      printLogMessageInNewLine: false,
+    });
+    logger.info("test 123");
+    expect(stdArray[3]).toBe("  \t");
+  });
+
+  /* display settings */
+
+  /* displayDateTime */
+  test("init logger: displayDateTime: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    logger.info("test 123");
+    expect(
+      doesLogContain(stdArray, new Date().getFullYear().toString())
+    ).toBeTruthy();
+  });
+
+  test("init logger: displayDateTime: true", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std, displayDateTime: true });
+    logger.info("test 123");
+    expect(
+      doesLogContain(stdArray, new Date().getFullYear().toString())
+    ).toBeTruthy();
+  });
+
+  test("init logger: displayDateTime: false", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std, displayDateTime: false });
+    logger.info("test 123");
+    expect(
+      doesLogContain(stdArray, new Date().getFullYear().toString())
+    ).toBeFalsy();
+  });
+
+  /* displayLogLevel */
+  test("init logger: displayLogLevel: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "INFO")).toBeTruthy();
+  });
+
+  test("init logger: displayLogLevel: true", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std, displayLogLevel: true });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "INFO")).toBeTruthy();
+  });
+
+  test("init logger: displayLogLevel: false", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std, displayLogLevel: false });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "INFO")).toBeFalsy();
+  });
+
+  /* displayInstanceName */
+  test("init logger: displayInstanceName: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std, instanceName: "name" });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "@name")).toBeFalsy();
+  });
+
+  test("init logger: displayInstanceName: true", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayInstanceName: true,
+      instanceName: "name",
+    });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "@name")).toBeTruthy();
+  });
+
+  test("init logger: displayInstanceName: false", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayInstanceName: false,
+      instanceName: "name",
+    });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "@name")).toBeFalsy();
+  });
+
+  /* displayLoggerName */
+  test("init logger: displayLoggerName: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std, name: "name" });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "name")).toBeTruthy();
+  });
+
+  test("init logger: displayLoggerName: true", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayLoggerName: true,
+      name: "name",
+    });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "name")).toBeTruthy();
+  });
+
+  test("init logger: displayLoggerName: false", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayLoggerName: false,
+      name: "name",
+    });
+    logger.info("test 123");
+    expect(doesLogContain(stdArray, "name")).toBeFalsy();
+  });
+
+  /* displayFunctionName */
+  test("init logger: displayFunctionName: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    function fnctn() {
+      logger.info("test 123");
+    }
+    fnctn();
+    expect(doesLogContain(stdArray, "fnctn")).toBeTruthy();
+  });
+
+  test("init logger: displayFunctionName: true", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayFunctionName: true,
+    });
+    function fnctn() {
+      logger.info("test 123");
+    }
+    fnctn();
+    expect(doesLogContain(stdArray, "fnctn")).toBeTruthy();
+  });
+
+  test("init logger: displayFunctionName: false", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayFunctionName: false,
+    });
+    function fnctn() {
+      logger.info("test 123");
+    }
+    fnctn();
+    expect(doesLogContain(stdArray, "fnctn")).toBeFalsy();
+  });
+
+  /* displayFilePath */
+  test("init logger: displayFilePath: default", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({ stdOut: std });
+    logger.info("test 123");
+
+    expect(doesLogContain(stdArray, "settings.test.ts")).toBeTruthy();
+  });
+
+  test("init logger: displayFilePath: hidden", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayFilePath: "hidden",
+    });
+    logger.info("test 123");
+
+    expect(doesLogContain(stdArray, "settings.test.ts")).toBeFalsy();
+  });
+
+  test("init logger: displayFilePath: displayAll", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayFilePath: "displayAll",
+    });
+    logger.info("test 123");
+
+    expect(doesLogContain(stdArray, "settings.test.ts")).toBeTruthy();
+  });
+
+  test("init logger: displayFilePath: hideNodeModulesOnly (not node_modules)", (): void => {
+    const stdArray: string[] = [];
+    const std: { write: (print: string) => void } = {
+      write: (print: string) => {
+        stdArray.push(print);
+      },
+    };
+    const logger: Logger = new Logger({
+      stdOut: std,
+      displayFilePath: "hideNodeModulesOnly",
+    });
+    logger.info("test 123");
+
+    expect(doesLogContain(stdArray, "settings.test.ts")).toBeTruthy();
+  });
+
+  /* Std */
   test("init logger: stdOut", (): void => {
     const std: { write: () => void } = { write: () => {} };
     const logger: Logger = new Logger({ stdOut: std });